requirements.txt
allure-pytest==2.13.5
allure-python-commons==2.13.5
pytest==8.3.3
selenium==4.24.0
seleniumbase==4.30.8

